# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Stars Decoded Natal Chart Report API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /horizon/preview:
    post:
      operationId: previewHorizon
      tags: [horizon]
      summary: What the entered birth time settles
      description: >
        Sweeps the birth-time band with the engine and returns, per fact,
        whether it holds and where it flips. Pure of the database: nothing is
        stored, no profile is read, no model is called. Small enough to poll on
        every debounced change of the form.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HorizonPreviewBody"
      responses:
        "200":
          description: The horizon facts across the band
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Horizon"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    HorizonFact:
      description: One fact the birth hour decides, swept across the band at two-minute steps.
      type: object
      properties:
        value:
          type: string
          description: The value at the centre time.
        holds:
          type: boolean
          description: True when the value never changed across the band.
        flipsAt:
          type: array
          description: Local times (HH:MM) inside the band at which the value changed.
          items:
            type: string
        values:
          type: array
          description: The sequence of values across the band.
          items:
            type: string
        holdsFrom:
          type: string
          description: Where the centre value's run within the birth day begins (HH:MM).
        holdsTo:
          type: string
          description: Where it ends (HH:MM, or 24:00).
      required: [value, holds, flipsAt, values, holdsFrom, holdsTo]
    Horizon:
      description: The horizon as a status, never a guess (ADR-33).
      type: object
      properties:
        status:
          type: string
          enum: [known, approximate, unknown]
        ascendant:
          $ref: "#/components/schemas/HorizonFact"
        midheaven:
          $ref: "#/components/schemas/HorizonFact"
        sect:
          $ref: "#/components/schemas/HorizonFact"
        moonSign:
          $ref: "#/components/schemas/HorizonFact"
        sunSign:
          $ref: "#/components/schemas/HorizonFact"
      required: [status, ascendant, midheaven, sect, moonSign, sunSign]
    HorizonPreviewBody:
      type: object
      properties:
        birthDate:
          type: string
        birthTime:
          type: string
          description: The band's centre, HH:MM.
        birthTimeWindowMinutes:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        timezone:
          type: string
          description: IANA zone name. Preferred over `timezoneOffset`.
        timezoneOffset:
          type: number
      required: [birthDate, birthTime, birthTimeWindowMinutes, latitude, longitude]
...